requirements.txt
amqp==5.1.1
asgiref==3.6.0
attrs==22.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==3.6.4.0
boto3==1.26.84
botocore==1.29.84
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
constantly==23.10.4
cryptography==39.0.2
daphne==4.0.0
defusedxml==0.8.0rc2
Django==4.1.7
django-birthday==0.1.4
django-cors-headers==3.14.0
django-countries==7.5.1
django-crontab==0.7.1
django-filter==23.2
django-multiselectfield==0.1.12
django-mysql==4.9.0
django-phonenumber-field==7.0.2
django-storages==1.13.2
django-subscribe==0.1.6
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.2
drf-nested-routers==0.93.5
drf-yasg==1.21.8
h5py==3.8.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
jmespath==1.0.1
kombu==5.2.4
mysql==0.0.3
mysqlclient==2.1.1
numpy==1.24.2
oauthlib==3.2.2
packaging==24.2
phonenumbers==8.13.7
pi==0.1.2
Pillow==9.4.0
pipp==0.0.1
prompt-toolkit==3.0.38
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.2.0
python-dateutil==2.8.2
python-decouple==3.8
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0.2
requests==2.28.2
requests-oauthlib==1.3.1
rest_condition==1.0.3
s3transfer==0.6.0
service-identity==24.1.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.3
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
validate_email==1.3
vine==5.0.0
wcwidth==0.2.6
zope.interface==6.1
